Complete deletion of LogsController
